Moved more components to archive.
