+++ src/components/old-components/ColorAndContrast.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { renderColorAndContrastBoxes } from '../utils';
+import { renderColorAndContrastBoxes } from '../../utils';
 
 interface ColorAndContrastProps {
     colors: string[];
